changed ints to doubles

+++ GraphingCalculator/src/Line.java
@@ -1,23 +1,23 @@
 import java.awt.*;
 
 public class Line {
-	protected int x1;
-	protected int y1;
-	protected int x2;
-	protected int y2;
+	protected double x1;
+	protected double y1;
+	protected double x2;
+	protected double y2;
 	protected Color color;
-	protected int stroke;
-	static final int DEFAULT_STROKE = 4;
+	protected double stroke;
+	static final double DEFAULT_STROKE = 4;
 	
-	public Line(int x1, int y1, int x2, int y2) {
+	public Line(double x1, double y1, double x2, double y2) {
 		this(x1, y1, x2, y2, DEFAULT_STROKE, Color.BLACK);
 	}
 	
-	public Line(int x1, int y1, int x2, int y2, int stroke) {
+	public Line(double x1, double y1, double x2, double y2, double stroke) {
 		this(x1, y1, x2, y2, stroke, Color.BLACK);
 	}
 	
-	public Line(int x1, int y1, int x2, int y2, int stroke, Color color) {
+	public Line(double x1, double y1, double x2, double y2, double stroke, Color color) {
 		this.x1 = x1;
 		this.y1 = y1;
 		this.x2 = x2;
